Split up test support files
