+++ tests/support/commands.ts
@@ -1,14 +1,4 @@
-import { expect } from '@playwright/test';
-import type { Page, Request, Locator, BrowserContext } from '@playwright/test';
-
-import type Swup from '../../src/Swup.js';
-
-declare global {
-	interface Window {
-		_swup: Swup;
-		measure: (key: string) => Promise<void>;
-	}
-}
+import { expect, Locator, Page, Request } from '@playwright/test';
 
 export function sleep(timeout = 0): Promise<void> {
   return new Promise((resolve) => setTimeout(() => resolve(undefined), timeout))
@@ -16,10 +6,6 @@ export function sleep(timeout = 0): Promise<void> {
 
 export function clickOnLink(page: Page, url: string, options?: Parameters<Page['click']>[1]) {
 	return page.click(`a[href="${url}"]`, options);
-}
-
-export function navigateWithSwup(page: Page, url: string, options?: Parameters<Swup['navigate']>[1]) {
-	return page.evaluate(({ url, options }) => window._swup.navigate(url, options), { url, options });
 }
 
 export async function expectToBeAt(page: Page, url: string, title?: string) {
@@ -42,35 +28,6 @@ export async function expectH2(page: Page, title: string) {
 	await expect(page.locator('h2')).toContainText(title);
 }
 
-export async function expectRequestHeaders(request: Request, headers: Record<string, string>) {
-	const expected = Object.fromEntries(
-		Object.entries(headers).map(([header, value]) => [header.toLowerCase(), value])
-	);
-	expect(request.headers()).toMatchObject(expected);
-}
-
-export async function expectToHaveCacheEntry(page: Page, url: string) {
-	const entry = await page.evaluate((url) => window._swup.cache.get(url), url);
-	expect(entry).toHaveProperty('url', url);
-}
-
-export async function expectToHaveCacheEntries(page: Page, urls: string[]) {
-	for (const url of urls) {
-		await expectToHaveCacheEntry(page, url);
-	}
-}
-
-export async function expectFullPageReload(page: Page, action: (page: Page) => Promise<void> | void, not: boolean = false) {
-	const origin = () => page.evaluate(() => window.performance.timeOrigin);
-	const before = await origin();
-	await action(page);
-	await expect(async () => expect(await origin()).not.toBe(before)).toPass();
-}
-
-export function expectSwupNavigation(page: Page, action: (page: Page) => Promise<void> | void) {
-	return expectFullPageReload(page, action, true);
-}
-
 export function expectToHaveText(locator: Locator, text: string) {
 	return expect(locator).toContainText(text);
 }
@@ -91,44 +48,22 @@ export function expectNotToHaveClasses(locator: Locator, classNames: string) {
 	return expectToHaveClasses(locator, classNames, true);
 }
 
-export async function expectAnimationDuration(page: Page, duration: number) {
-	const tolerance = duration ? 0.25 : 0; // 25% plus/minus
-	const expectedRange: [number, number] = [
-		duration * (1 - tolerance),
-		duration * (1 + tolerance)
-	];
-	let timing = { start: 0, end: 0, outStart: 0, outEnd: 0, inStart: 0, inEnd: 0 };
+export async function expectPageReload(page: Page, action: (page: Page) => Promise<void> | void, not: boolean = false) {
+	const origin = () => page.evaluate(() => window.performance.timeOrigin);
+	const before = await origin();
+	await action(page);
+	await expect(async () => expect(await origin()).not.toBe(before)).toPass();
+}
 
-	// Make sure we're ready to animate
-	await page.waitForSelector('html.swup-enabled');
-	// Make sure we're not disabling animations
-	await page.emulateMedia({ reducedMotion: null });
+export function expectNoPageReload(page: Page, action: (page: Page) => Promise<void> | void) {
+	return expectPageReload(page, action, true);
+}
 
-	await page.evaluate(() => {
-		window.data = {};
-		const measure = (key, value?: number) => window.data[key] = value ?? performance.now();
-		window._swup.hooks.on('visit:start', () => measure('start'));
-		window._swup.hooks.on('visit:end', () => measure('end'));
-		window._swup.hooks.on('animation:out:start', () => measure('outStart'));
-		window._swup.hooks.on('animation:out:end', () => measure('outEnd'));
-		window._swup.hooks.on('animation:in:start', () => measure('inStart'));
-		window._swup.hooks.on('animation:in:end', () => measure('inEnd'));
-		window._swup.hooks.on('animation:skip', () => measure('inStart', 0));
-		window._swup.hooks.on('animation:skip', () => measure('inEnd', 0));
-		window._swup.hooks.on('animation:skip', () => measure('outStart', 0));
-		window._swup.hooks.on('animation:skip', () => measure('outEnd', 0));
-	});
-
-	await navigateWithSwup(page, page.url());
-	await page.waitForFunction(() => window.data.end > 0);
-	timing = await page.evaluate(() => window.data);
-
-	const outDuration = timing.outEnd - timing.outStart;
-	const inDuration = timing.inEnd - timing.inStart;
-	expect(outDuration).toBeGreaterThanOrEqual(expectedRange[0]);
-	expect(outDuration).toBeLessThanOrEqual(expectedRange[1]);
-	expect(inDuration).toBeGreaterThanOrEqual(expectedRange[0]);
-	expect(inDuration).toBeLessThanOrEqual(expectedRange[1]);
+export async function expectRequestHeaders(request: Request, headers: Record<string, string>) {
+	const expected = Object.fromEntries(
+		Object.entries(headers).map(([header, value]) => [header.toLowerCase(), value])
+	);
+	expect(request.headers()).toMatchObject(expected);
 }
 
 export async function delayRequest(page: Page, url: string, timeout: number) {
@@ -148,9 +83,4 @@ export async function expectScrollPosition(page: Page, y: number) {
 			await page.evaluate(() => window.scrollY)
 		).toEqual(y)
 	).toPass();
-}
-
-export async function pushHistoryState(page: Page, url: string, data: Record<string, unknown> = {}) {
-	const state = { url, random: Math.random(), source: 'swup', ...data };
-	await page.evaluate(({ url, state }) => window.history.pushState(state, '', url), { url, state });
 }
